Add Mimir algorithm (#8)
* Fix Olken and refactor object methods to use double underscores
* Change default line limit to 80 characters
* Refactor to remove compiler warnings. Also lowered error level to get rid of -Wpadding and -Wunsuffixed-...
* Add histogram plotter
* Add Zipfian runtime complexity note
* Compile static libraries
* Add performance test
* Make it compatible with gprof

Mimir seems to take a while to iterate through the oldest bucket in the histogram. Weird bc it shouldn't be called that often. I looked at the distribution and it may be that it isn't skewed enough.

Looking at gprof, it seems that the stacker aging policy takes most of the time because it calls the aging function so often. I'd have to look into this.