Test speed of bucketed sampling method -- completely inaccurate (#39)
* Improve starting performance of per-bucket sampling

I had to lower the initial threshold. There is some dead code lying around. I thought the main bottleneck was needing to access the hash table first, but it's not.

I also merged the lookup and put functions into a "try-put" function that attempts a put and succeeds if the new hash value is lower. I had to inline the function (i.e. make it `static inline`) to get good performance.

* Refactor Histogram to use bins

* Add fixed rate SHARDS to QuickMRC