Debug AET accuracy (#49)
It turns out AET has good accuracy for the first cliff but poor accuracy on cliffs thereafter. It seems that AET performs well when the miss rate is high.

In fact, AET does not necessarily perform well when the reuse time is similar to the reuse distance. Consider the following trace:

> 1, 2, 3, 1, 2, 3, 1, 2, 3, 4, 5, 6, 1, 2, 3, 4, 5, 6

This will yield a reuse time and distance histogram of:

> {3: 6, 6: 6, \infty: 6}

However, the resulting MRC from Olken versus AET is different, with AET being compressed.