Separate hash from key/value in evicting hash table
We predict this allows closer to the whole hash array to reside in
the L1 cache, thereby improving performance. If we want to parallelize
this, we will need to revert this so that we can atomically update each
bucket and only modify a single cache line.

The original reason for this was because I thought it would improve the
global threshold scan. I don't believe it significantly changed the
results, however.